Added UserUpdateProcessor tests
