+++ service-api/src/main/java/greencity/service/ubs/TelegramCommandsService.java
@@ -1,0 +1,8 @@
+package greencity.service.ubs;
+
+import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
+import org.telegram.telegrambots.meta.api.objects.Message;
+
+public interface TelegramCommandsService {
+    SendMessage processCommand(Message message);
+}
